Fix Good OverHeating + fix Ai in formation in sloap

+++ SquadLeader/Source/SquadLeader/AI/AIClassesController/SupportAISquadController.cpp
@@ -14,8 +14,8 @@ ASupportAISquadController::ASupportAISquadController() {
 }
 
 void ASupportAISquadController::CheckIfLaunchHeal() {
-	int TotalHealth = 0;
-	int TotalMaxHealth = 0;
+	float TotalHealth = 0;
+	float TotalMaxHealth = 0;
 	float RatioHealth = 0;
 
 	bool IsAnAllyHealthVeryLow = false;
